LCTV contribution!! Mummi contribution (for infinities...) !! catherinacortega contribution (for weeks)!! danielv contribution!! xmetrix streamer and chat contribution (long time and this morning long)!! dchem chat contribution!! figjeti and others pals (contribution...) contribution...!!
Thank you Carlos!!

Thank you Mummi!!

Thank you Carol!!

from inha-macos!!
i7 2600 16GB 1333MHz DDR3 2x500GB HDD
at /macos/ folder at inha-macos at veriverse!!

+++ appseed/aura/aura/os/macos/macos_time.cpp
@@ -46,7 +46,7 @@ uint64_t get_nanos()
     ts.tv_sec = mts.tv_sec;
     ts.tv_nsec = mts.tv_nsec;
     
-    return ts.tv_sec * 1000 * 1000 * 1000 + ts.tv_nsec;
+    return (uint64_t) ts.tv_sec * 1000 * 1000 * 1000 + (uint64_t) ts.tv_nsec;
     
 }
